Resolved conflicts in JUnit tests cases
