+++ src/main/java/com/searchservice/app/rest/ManageTableResource.java
@@ -65,7 +65,7 @@ public class ManageTableResource {
 
 	@GetMapping("/")
 	@Operation(summary = "GET ALL THE TABLES FOR THE GIVEN TENANT ID.", security = @SecurityRequirement(name = "bearerAuth"))
-	@PreAuthorize(value = "@keycloakUserPermission.isWritePermissionEnabled()")
+	@PreAuthorize(value = "@keycloakUserPermission.isViewPermissionEnabled()")
 	public ResponseEntity<Response> getTables(@RequestParam int tenantId) {
 
 		Response getListItemsResponseDTO = manageTableServicePort.getTables(tenantId);
@@ -87,6 +87,7 @@ public class ManageTableResource {
 	
 	@GetMapping("/all-tables")
 	@Operation(summary = "GET ALL THE TABLES FROM THE SERVER.", security = @SecurityRequirement(name = "bearerAuth"))
+	@PreAuthorize(value = "@keycloakUserPermission.isViewPermissionEnabled()")
 	public ResponseEntity<Response> getALLTables(@RequestParam(defaultValue = "1") int pageNumber,
             @RequestParam(defaultValue = "5") int pageSize) {
 
@@ -108,6 +109,7 @@ public class ManageTableResource {
 	
 	@GetMapping("/deletion/all-tables")
 	@Operation(summary = "GET ALL THE TABLES UNDER DELETION.", security = @SecurityRequirement(name = "bearerAuth"))
+	@PreAuthorize(value = "@keycloakUserPermission.isViewPermissionEnabled()")
 	public ResponseEntity<Response> getALLTablesUnderDeletion(@RequestParam(defaultValue = "1") int pageNumber,
             @RequestParam(defaultValue = "5") int pageSize) {
 
@@ -129,6 +131,7 @@ public class ManageTableResource {
 
 	@GetMapping("/{tableName}")
 	@Operation(summary = "GET SCHEMA OF A TABLE.", security = @SecurityRequirement(name = "bearerAuth"))
+	@PreAuthorize(value = "@keycloakUserPermission.isViewPermissionEnabled()")
 	public ResponseEntity<TableSchema> getTable(@RequestParam int tenantId, @PathVariable String tableName) {
 
 		if (tableDeleteServicePort.isTableUnderDeletion(tableName + "_"+ tenantId)) {
@@ -152,6 +155,7 @@ public class ManageTableResource {
 
 	@PostMapping("/")
 	@Operation(summary = "CREATE A TABLE UNDER THE GIVEN TENANT ID.", security = @SecurityRequirement(name = "bearerAuth"))
+	@PreAuthorize(value = "@keycloakUserPermission.isCreatePermissionEnabled()")
 	public ResponseEntity<Response> createTable(@RequestParam int tenantId, @RequestBody CreateTable createTableDTO) {
 
 		if (manageTableServicePort.checkIfTableNameisValid(createTableDTO.getTableName())) {
@@ -187,6 +191,7 @@ public class ManageTableResource {
 		
 	@DeleteMapping("/{tableName}")
 	@Operation(summary = "DELETE A TABLE (SOFT DELETE).", security = @SecurityRequirement(name = "bearerAuth"))
+	@PreAuthorize(value = "@keycloakUserPermission.isDeletePermissionEnabled()")
 	public ResponseEntity<Response> deleteTable(@RequestParam int tenantId, @PathVariable String tableName) {
 		tableName = tableName + "_" + tenantId;
 		if (!tableDeleteServicePort.isTableUnderDeletion(tableName)) {
@@ -212,6 +217,7 @@ public class ManageTableResource {
 
 	@PutMapping("/restore/{tableName}")
 	@Operation(summary = "RESTORE A DELETED TABLE.", security = @SecurityRequirement(name = "bearerAuth"))
+	@PreAuthorize(value = "@keycloakUserPermission.isEditPermissionEnabled()")
 	public ResponseEntity<Response> restoreTable(@RequestParam int tenantId, @PathVariable String tableName) {
 		String tableNameForMessage = tableName;
 		tableName = tableName + "_" + tenantId;
@@ -233,6 +239,7 @@ public class ManageTableResource {
 
 	@PutMapping("/{tableName}")
 	@Operation(summary = "REPLACE SCHEMA OF AN EXISTING TABLE.", security = @SecurityRequirement(name = "bearerAuth"))
+	@PreAuthorize(value = "@keycloakUserPermission.isEditPermissionEnabled()")
 	public ResponseEntity<Response> updateTableSchema(@RequestParam int tenantId, @PathVariable String tableName,
 			@RequestBody ManageTable newTableSchemaDTO) {
